Added framework for frequency method testing
This is a start of a framework for testing of different frequency methods. 
frequency_methods.py contains functions for frequency calculation. The implementations are merely a start and improvements can be made, especially for the phase separation method. 
test_signal_generation.py contains classes and functions for generation of test waveforms and input of xls data from file.
testbench.py contains some testing code. This can be expanded to include automated testing of the performance of different methods. 
Reference signal for testing.xlsx contains a signal provided by Roger Bergeron in xls format.